refactor: StrategiesPage 컴포넌트 분리 — 641줄 → 155줄 (#92)
StrategyCard, ParamsEditor, SimulationResult를 components/strategies/로 추출.
미사용 CoinStrategyTab, EditField 제거.
