+++ admin/src/pages/StrategiesPage.tsx
@@ -1,13 +1,13 @@
-import { useEffect, useState, useCallback } from "react";
-import { getStrategies, getActivationHistory, activateStrategy, deactivateStrategy, updateStrategyParams, getStrategySimulation } from "../api/strategies";
-import { getAllCoinStrategies, updateCoinStrategy } from "../api/coinStrategy";
+import { useEffect, useState } from "react";
+import { getStrategies, getActivationHistory, activateStrategy, deactivateStrategy } from "../api/strategies";
+import { getAllCoinStrategies } from "../api/coinStrategy";
 import type { CoinStrategyConfig } from "../api/coinStrategy";
-import type { Strategy, StrategyActivation, StrategySimulation } from "../types/strategies";
+import type { Strategy, StrategyActivation } from "../types/strategies";
 import ConfirmDialog from "../components/ConfirmDialog";
-import { formatPercent, formatDateTime, formatKRW } from "../utils/format";
-import { getParamDesc } from "../utils/paramDescriptions";
 import { getMarketStateKR } from "../utils/indicatorDescriptions";
 import client from "../api/client";
+import StrategyCard from "../components/strategies/StrategyCard";
+import ParamsEditor from "../components/strategies/ParamsEditor";
 
 const MARKET_SECTIONS = [
   { state: "sideways", label: "횡보장 전략", emoji: "➡️", desc: "변동이 적은 박스권에서 유리" },
@@ -15,13 +15,7 @@ const MARKET_SECTIONS = [
   { state: "bearish", label: "하락장 전략", emoji: "📉", desc: "하락 추세에서 유리" },
 ] as const;
 
-const CATEGORY_INFO: Record<string, { label: string; emoji: string; desc: string }> = {
-  core: { label: "대형코인", emoji: "🏦", desc: "BTC, ETH, XRP — 변동 적음, 보수적 운용" },
-  alt: { label: "알트코인", emoji: "🚀", desc: "자동 선별 알트코인 — 변동 큼, 공격적 운용" },
-};
-
 export default function StrategiesPage() {
-  const [tab, setTab] = useState<"strategies" | "coin-config">("strategies");
   const [strategies, setStrategies] = useState<Strategy[]>([]);
   const [activations, setActivations] = useState<StrategyActivation[]>([]);
   const [coinConfigs, setCoinConfigs] = useState<CoinStrategyConfig[]>([]);
@@ -67,7 +61,6 @@ export default function StrategiesPage() {
 
   if (loading) return <div className="loading">로딩 중...</div>;
 
-  const activeStrategies = strategies.filter((s) => s.is_active);
   const hasSwitching = strategies.some((s) => s.status === "shutting_down");
 
   return (
@@ -171,471 +164,6 @@ export default function StrategiesPage() {
           onSaved={fetchData}
         />
       )}
-    </div>
-  );
-}
-
-
-// ── 코인별 전략 설정 탭 ──
-
-function CoinStrategyTab({ configs, strategies, onSaved }: {
-  configs: CoinStrategyConfig[];
-  strategies: Strategy[];
-  onSaved: () => void;
-}) {
-  const [saving, setSaving] = useState<string | null>(null);
-  const [editState, setEditState] = useState<Record<string, {
-    strategy_name: string;
-    stop_loss_pct: string;
-    trailing_stop_pct: string;
-    position_size_pct: string;
-    strategy_params_json: string;
-  }>>({});
-
-  useEffect(() => {
-    const state: typeof editState = {};
-    configs.forEach((c) => {
-      state[c.category] = {
-        strategy_name: c.strategy_name,
-        stop_loss_pct: String(c.stop_loss_pct),
-        trailing_stop_pct: String(c.trailing_stop_pct),
-        position_size_pct: String(c.position_size_pct),
-        strategy_params_json: c.strategy_params_json || "{}",
-      };
-    });
-    setEditState(state);
-  }, [configs]);
-
-  const handleSave = async (category: string) => {
-    const edit = editState[category];
-    if (!edit) return;
-    setSaving(category);
-    try {
-      await updateCoinStrategy(category, {
-        strategy_name: edit.strategy_name,
-        stop_loss_pct: parseFloat(edit.stop_loss_pct),
-        trailing_stop_pct: parseFloat(edit.trailing_stop_pct),
-        position_size_pct: parseFloat(edit.position_size_pct),
-        strategy_params_json: edit.strategy_params_json,
-      });
-      onSaved();
-    } finally {
-      setSaving(null);
-    }
-  };
-
-  const strategyNames = strategies.map((s) => s.name);
-
-  return (
-    <div>
-      {configs.map((cfg) => {
-        const info = CATEGORY_INFO[cfg.category] || { label: cfg.category, emoji: "📦", desc: "" };
-        const edit = editState[cfg.category];
-        if (!edit) return null;
-
-        return (
-          <div key={cfg.category} className="card" style={{ marginBottom: 20 }}>
-            <div className="card-title">
-              <span style={{ marginRight: 8 }}>{info.emoji}</span>
-              {info.label}
-              <span style={{ fontSize: 12, color: "var(--text-muted)", marginLeft: 8 }}>{info.desc}</span>
-            </div>
-
-            <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 16 }}>
-              {/* 전략 선택 */}
-              <div>
-                <div style={{ fontSize: 12, color: "var(--text-muted)", marginBottom: 4 }}>매매 전략</div>
-                <select
-                  value={edit.strategy_name}
-                  onChange={(e) => setEditState((prev) => ({ ...prev, [cfg.category]: { ...edit, strategy_name: e.target.value } }))}
-                  style={{
-                    width: "100%", padding: "8px", borderRadius: 6,
-                    border: "1px solid var(--border)", background: "var(--bg-secondary)",
-                    color: "var(--text-primary)", fontSize: 13,
-                  }}
-                >
-                  {strategyNames.map((name) => (
-                    <option key={name} value={name}>{name}</option>
-                  ))}
-                </select>
-              </div>
-
-              {/* 포지션 크기 */}
-              <EditField label="포지션 크기 (%)" caption="잔고 대비 투자 비율" value={edit.position_size_pct}
-                onChange={(v) => setEditState((prev) => ({ ...prev, [cfg.category]: { ...edit, position_size_pct: v } }))} />
-
-              {/* 손절 */}
-              <EditField label="손절률 (%)" caption="매수가 대비 이 비율 하락 시 자동 매도" value={edit.stop_loss_pct}
-                onChange={(v) => setEditState((prev) => ({ ...prev, [cfg.category]: { ...edit, stop_loss_pct: v } }))} />
-
-              {/* 트레일링 */}
-              <EditField label="트레일링 스탑 (%)" caption="최고가 대비 이 비율 하락 시 자동 매도" value={edit.trailing_stop_pct}
-                onChange={(v) => setEditState((prev) => ({ ...prev, [cfg.category]: { ...edit, trailing_stop_pct: v } }))} />
-
-              {/* 전략 파라미터 */}
-              <div style={{ gridColumn: "1 / -1" }}>
-                <div style={{ fontSize: 12, color: "var(--text-muted)", marginBottom: 4 }}>전략 파라미터 (JSON)</div>
-                <input
-                  value={edit.strategy_params_json}
-                  onChange={(e) => setEditState((prev) => ({ ...prev, [cfg.category]: { ...edit, strategy_params_json: e.target.value } }))}
-                  style={{
-                    width: "100%", padding: "8px", borderRadius: 6, fontFamily: "monospace",
-                    border: "1px solid var(--border)", background: "var(--bg-secondary)",
-                    color: "var(--text-primary)", fontSize: 12,
-                  }}
-                />
-                <div style={{ fontSize: 11, color: "var(--text-muted)", marginTop: 2 }}>
-                  {cfg.description}
-                </div>
-              </div>
-            </div>
-
-            <div style={{ marginTop: 16, display: "flex", justifyContent: "flex-end" }}>
-              <button
-                onClick={() => handleSave(cfg.category)}
-                disabled={saving === cfg.category}
-                className="btn btn-primary btn-sm"
-                style={{ opacity: saving === cfg.category ? 0.6 : 1 }}
-              >
-                {saving === cfg.category ? "저장 중..." : "설정 저장"}
-              </button>
-            </div>
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-function EditField({ label, caption, value, onChange }: {
-  label: string; caption: string; value: string; onChange: (v: string) => void;
-}) {
-  return (
-    <div>
-      <div style={{ fontSize: 12, color: "var(--text-muted)", marginBottom: 4 }}>{label}</div>
-      <input
-        type="number" step="0.5" value={value}
-        onChange={(e) => onChange(e.target.value)}
-        style={{
-          width: "100%", padding: "8px", borderRadius: 6,
-          border: "1px solid var(--border)", background: "var(--bg-secondary)",
-          color: "var(--text-primary)", fontSize: 13,
-        }}
-      />
-      <div style={{ fontSize: 11, color: "var(--text-muted)", marginTop: 2 }}>{caption}</div>
-    </div>
-  );
-}
-
-
-// ── 전략 카드 ──
-
-function StrategyCard({ strategy: s, hasSwitching, activeCoins, onEdit, onActivate, onDeactivate }: {
-  strategy: Strategy;
-  hasSwitching: boolean;
-  activeCoins: string[];
-  onEdit: () => void;
-  onActivate: () => void;
-  onDeactivate: () => void;
-}) {
-  return (
-    <div
-      className="strategy-card"
-      onClick={onEdit}
-      style={{ cursor: "pointer" }}
-    >
-      <div style={{ display: "flex", justifyContent: "space-between", alignItems: "flex-start" }}>
-        <div>
-          <h3>{s.display_name}</h3>
-          <div style={{ display: "flex", gap: 6, marginBottom: 4 }}>
-            <span className="badge badge-purple">{s.category}</span>
-            <span className="badge badge-blue">{s.difficulty}</span>
-          </div>
-        </div>
-        {activeCoins.length > 0 ? (
-          <div style={{ display: "flex", gap: 4, flexWrap: "wrap" }}>
-            {activeCoins.map((c) => (
-              <span key={c} className="badge badge-green" style={{ fontSize: 10 }}>{c}</span>
-            ))}
-          </div>
-        ) : (
-          <span className="badge badge-red" style={{ fontSize: 10 }}>미사용</span>
-        )}
-      </div>
-      <p className="strategy-desc">{s.description}</p>
-      <div className="strategy-stats">
-        <div className="strategy-stat-item">
-          <div className="stat-label">총 거래</div>
-          <div className="stat-value">{s.stats?.total_trades ?? 0}</div>
-        </div>
-        <div className="strategy-stat-item">
-          <div className="stat-label">승률</div>
-          <div className="stat-value">{s.stats?.total_trades > 0 && s.stats?.win_rate != null ? `${s.stats.win_rate.toFixed(1)}%` : "N/A"}</div>
-        </div>
-        <div className="strategy-stat-item">
-          <div className="stat-label">평균 수익률</div>
-          <div className={`stat-value ${(s.stats?.avg_profit_pct ?? 0) >= 0 ? "positive" : "negative"}`}>
-            {s.stats?.avg_profit_pct != null ? formatPercent(s.stats.avg_profit_pct) : "-"}
-          </div>
-        </div>
-      </div>
-      <div style={{ display: "flex", gap: 8 }}>
-        <span className="badge badge-blue">{s.timeframe}</span>
-      </div>
-    </div>
-  );
-}
-
-
-// ── 파라미터 편집 모달 ──
-
-function ParamsEditor({ strategy, onClose, onSaved }: {
-  strategy: Strategy;
-  onClose: () => void;
-  onSaved: () => void;
-}) {
-  const currentParams = JSON.parse(strategy.default_params_json || "{}");
-  const [editParams, setEditParams] = useState<Record<string, string>>({});
-  const [simulation, setSimulation] = useState<StrategySimulation | null>(null);
-  const [saving, setSaving] = useState(false);
-  const [simLoading, setSimLoading] = useState(false);
-
-  useEffect(() => {
-    const initial: Record<string, string> = {};
-    Object.entries(currentParams).forEach(([k, v]) => {
-      initial[k] = String(v);
-    });
-    setEditParams(initial);
-  }, [strategy.name]);
-
-  const hasChanges = Object.keys(editParams).some(
-    (k) => editParams[k] !== String(currentParams[k])
-  );
-
-  const runSimulation = useCallback(async () => {
-    setSimLoading(true);
-    try {
-      const newParams: Record<string, number> = {};
-      Object.entries(editParams).forEach(([k, v]) => {
-        newParams[k] = parseFloat(v) || 0;
-      });
-      const result = await getStrategySimulation(strategy.name, JSON.stringify(newParams));
-      setSimulation(result);
-    } catch {
-      // ignore
-    } finally {
-      setSimLoading(false);
-    }
-  }, [editParams, strategy.name]);
-
-  // 파라미터 변경 시 자동 시뮬레이션
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      runSimulation();
-    }, 300);
-    return () => clearTimeout(timer);
-  }, [editParams, runSimulation]);
-
-  const handleSave = async () => {
-    setSaving(true);
-    try {
-      const newParams: Record<string, number> = {};
-      Object.entries(editParams).forEach(([k, v]) => {
-        newParams[k] = parseFloat(v) || 0;
-      });
-      await updateStrategyParams(strategy.name, JSON.stringify(newParams));
-      onSaved();
-      onClose();
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const sim = simulation?.simulation || {};
-
-  return (
-    <div className="modal-overlay" onClick={onClose}>
-      <div className="modal-content" onClick={(e) => e.stopPropagation()} style={{ maxWidth: 700, width: "90%" }}>
-        <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", marginBottom: 20 }}>
-          <div>
-            <h3 style={{ margin: 0 }}>{strategy.display_name} — 파라미터 설정</h3>
-            <div style={{ fontSize: 12, color: "var(--text-muted)", marginTop: 4 }}>{strategy.description}</div>
-          </div>
-          <button onClick={onClose} style={{ background: "none", border: "none", color: "var(--text-secondary)", fontSize: 18, cursor: "pointer" }}>x</button>
-        </div>
-
-        {/* 파라미터 편집 — 설명 포함 */}
-        <div style={{ marginBottom: 20 }}>
-          <div style={{ display: "grid", gridTemplateColumns: "1fr auto auto", gap: 0, fontSize: 12, fontWeight: 600, color: "var(--text-muted)", padding: "0 0 8px", borderBottom: "1px solid var(--border)" }}>
-            <span>파라미터</span>
-            <span style={{ width: 80, textAlign: "center" }}>현재값</span>
-            <span style={{ width: 90, textAlign: "center", color: "#4a9eff" }}>변경값</span>
-          </div>
-          {Object.entries(editParams).map(([key, value]) => {
-            const desc = getParamDesc(strategy.name, key);
-            const changed = value !== String(currentParams[key]);
-            return (
-              <div key={key} style={{ padding: "12px 0", borderBottom: "1px solid var(--border)" }}>
-                <div style={{ display: "grid", gridTemplateColumns: "1fr auto auto", gap: 8, alignItems: "center" }}>
-                  <div>
-                    <div style={{ fontSize: 14, fontWeight: 600 }}>
-                      {desc.label}
-                      {desc.unit && <span style={{ fontSize: 11, color: "var(--text-muted)", marginLeft: 4 }}>({desc.unit})</span>}
-                    </div>
-                  </div>
-                  <div style={{ width: 80, textAlign: "center", fontSize: 14, color: changed ? "var(--text-muted)" : "var(--text-primary)", textDecoration: changed ? "line-through" : "none" }}>
-                    {String(currentParams[key])}
-                  </div>
-                  <input
-                    type="number"
-                    step={desc.step || 0.05}
-                    min={desc.min}
-                    max={desc.max}
-                    value={value}
-                    onChange={(e) => setEditParams((prev) => ({ ...prev, [key]: e.target.value }))}
-                    style={{
-                      width: 90,
-                      padding: "6px 8px",
-                      borderRadius: 6,
-                      border: changed ? "2px solid #4a9eff" : "1px solid var(--border)",
-                      background: "var(--bg-secondary)",
-                      color: "var(--text-primary)",
-                      fontSize: 14,
-                      textAlign: "right",
-                      fontWeight: 600,
-                    }}
-                  />
-                </div>
-                <div style={{ fontSize: 12, color: "var(--text-muted)", marginTop: 4 }}>{desc.description}</div>
-                {desc.tip && (
-                  <div style={{ fontSize: 11, color: "#4a9eff", marginTop: 2 }}>
-                    {desc.tip}
-                  </div>
-                )}
-              </div>
-            );
-          })}
-        </div>
-
-        {/* 시뮬레이션 결과 */}
-        {simLoading ? (
-          <div style={{ textAlign: "center", padding: 16, color: "var(--text-muted)" }}>시뮬레이션 중...</div>
-        ) : Object.keys(sim).length > 0 ? (
-          <div style={{ background: "var(--bg-secondary)", borderRadius: 8, padding: 16, marginBottom: 16 }}>
-            <div style={{ fontSize: 13, fontWeight: 600, marginBottom: 12 }}>
-              현재 시장 기반 시뮬레이션
-            </div>
-            <SimulationResult strategyName={strategy.name} sim={sim} hasChanges={hasChanges} />
-          </div>
-        ) : null}
-
-        {/* 저장 버튼 */}
-        <div style={{ display: "flex", justifyContent: "flex-end", gap: 8 }}>
-          <button onClick={onClose} className="btn btn-sm" style={{ background: "#2a2d3e" }}>취소</button>
-          <button
-            onClick={handleSave}
-            disabled={!hasChanges || saving}
-            className="btn btn-primary btn-sm"
-            style={{ opacity: hasChanges ? 1 : 0.4 }}
-          >
-            {saving ? "저장 중..." : "파라미터 적용"}
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-
-function SimulationResult({ strategyName, sim, hasChanges }: {
-  strategyName: string;
-  sim: Record<string, number | string | null>;
-  hasChanges: boolean;
-}) {
-  if (strategyName === "bollinger_bands" || strategyName === "bollinger_squeeze") {
-    return (
-      <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 12, fontSize: 13 }}>
-        <SimRow label="현재 BTC" value={formatKRW(Number(sim.current_price))} />
-        <SimRow label="MA(20)" value={formatKRW(Number(sim.ma_20))} />
-        <SimRow label="현재 상단" value={formatKRW(Number(sim.current_upper))} changed={hasChanges} />
-        <SimRow label="변경 상단" value={formatKRW(Number(sim.new_upper))} highlight={hasChanges} />
-        <SimRow label="현재 하단" value={formatKRW(Number(sim.current_lower))} changed={hasChanges} />
-        <SimRow label="변경 하단" value={formatKRW(Number(sim.new_lower))} highlight={hasChanges} />
-        <SimRow label="현재 밴드폭" value={formatKRW(Number(sim.current_band_width))} changed={hasChanges} />
-        <SimRow label="변경 밴드폭" value={formatKRW(Number(sim.new_band_width))} highlight={hasChanges} />
-        <SimRow
-          label="현재 하단까지"
-          value={`${formatKRW(Number(sim.current_distance_to_lower))} (${sim.current_distance_to_lower_pct}%)`}
-          changed={hasChanges}
-        />
-        <SimRow
-          label="변경 하단까지"
-          value={`${formatKRW(Number(sim.new_distance_to_lower))} (${sim.new_distance_to_lower_pct}%)`}
-          highlight={hasChanges}
-        />
-      </div>
-    );
-  }
-
-  if (strategyName === "volatility_breakout") {
-    return (
-      <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 12, fontSize: 13 }}>
-        <SimRow label="현재 BTC" value={formatKRW(Number(sim.current_price))} />
-        <SimRow label="금일 시가" value={formatKRW(Number(sim.today_open))} />
-        <SimRow label="전일 변동폭" value={formatKRW(Number(sim.price_range))} />
-        <div />
-        <SimRow label="현재 돌파선" value={formatKRW(Number(sim.current_breakout))} changed={hasChanges} />
-        <SimRow label="변경 돌파선" value={formatKRW(Number(sim.new_breakout))} highlight={hasChanges} />
-        <SimRow label="현재 거리" value={formatKRW(Number(sim.current_distance))} changed={hasChanges} />
-        <SimRow label="변경 거리" value={formatKRW(Number(sim.new_distance))} highlight={hasChanges} />
-      </div>
-    );
-  }
-
-  if (strategyName === "rsi_mean_reversion") {
-    return (
-      <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 12, fontSize: 13 }}>
-        <SimRow label="현재 RSI" value={String(sim.current_rsi)} />
-        <div />
-        <SimRow label="현재 과매도" value={String(sim.current_oversold)} changed={hasChanges} />
-        <SimRow label="변경 과매도" value={String(sim.new_oversold)} highlight={hasChanges} />
-        <SimRow label="현재 과매수" value={String(sim.current_overbought)} changed={hasChanges} />
-        <SimRow label="변경 과매수" value={String(sim.new_overbought)} highlight={hasChanges} />
-        <SimRow label="현재 매수까지" value={`RSI ${sim.current_buy_distance} 남음`} changed={hasChanges} />
-        <SimRow label="변경 매수까지" value={`RSI ${sim.new_buy_distance} 남음`} highlight={hasChanges} />
-      </div>
-    );
-  }
-
-  // 기본: key-value 나열
-  return (
-    <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 8, fontSize: 13 }}>
-      {Object.entries(sim).map(([key, value]) => (
-        <SimRow key={key} label={key} value={value != null ? String(typeof value === "number" ? value.toLocaleString() : value) : "-"} />
-      ))}
-    </div>
-  );
-}
-
-
-function SimRow({ label, value, changed, highlight }: {
-  label: string;
-  value: string;
-  changed?: boolean;
-  highlight?: boolean;
-}) {
-  return (
-    <div style={{
-      display: "flex",
-      justifyContent: "space-between",
-      padding: "4px 8px",
-      borderRadius: 4,
-      background: highlight ? "rgba(74, 158, 255, 0.1)" : "transparent",
-      opacity: changed ? 0.5 : 1,
-      textDecoration: changed ? "line-through" : "none",
-    }}>
-      <span style={{ color: "var(--text-muted)" }}>{label}</span>
-      <span style={{ fontWeight: 600, color: highlight ? "#4a9eff" : "var(--text-primary)" }}>{value}</span>
     </div>
   );
 }
